+++ backend/src/main/java/com/yagubogu/stat/repository/VictoryFairyRankingRepositoryCustom.java
@@ -12,7 +12,7 @@ public interface VictoryFairyRankingRepositoryCustom {
 
     Optional<VictoryFairyRankParam> findByMemberAndTeamFilterAndYear(Member member, TeamFilter teamFilter, int year);
 
-    Optional<Integer> findRankWithinTeamByMemberAndYear(Member member, int year);
+    Optional<Long> findRankWithinTeamByMemberAndYear(Member member, int year);
 
     void batchUpdate(List<UpdateDto> updates, int batchSize);
